remove obsolete error defs
