+++ errors/errors.go
@@ -74,6 +74,19 @@ var (
 	ErrPanic = Register(111222, "panic")
 )
 
+// TMError is the tendermint abci return type with stack trace
+type TMError interface {
+	stackTracer
+	ABCICode() uint32
+	ABCILog() string
+}
+
+// stackTracer from pkg/errors
+type stackTracer interface {
+	error
+	StackTrace() errors.StackTrace
+}
+
 // Register returns an error instance that should be used as the base for
 // creating error instances during runtime.
 //
@@ -320,4 +333,45 @@ func (e *wrappedError) Format(s fmt.State, verb rune) {
 		fmt.Fprintf(s, e.ABCILog())
 		writeSimpleFrame(s, stack[0])
 	}
+}
+
+// HasErrorCode checks if this error would return the named error code
+func HasErrorCode(err error, code uint32) bool {
+	if tm, ok := err.(coder); ok {
+		return tm.ABCICode() == code
+	}
+	return code == ErrInternal.ABCICode()
+}
+
+// NormalizePanic converts a panic into a redacted error
+//
+// We want the whole stack trace for logging
+// but should show nothing over the ABCI interface....
+func NormalizePanic(p interface{}) error {
+	// TODO, handle this better??? for stack traces
+	// if err, isErr := p.(error); isErr {
+	// 	return Wrap(err, "normalized panic")
+	// }
+	return ErrPanic.Newf("%v", p)
+}
+
+// Redact will replace all panic errors with a generic message
+func Redact(err error) error {
+	if HasErrorCode(err, ErrPanic.code) {
+		return ErrInternal
+	}
+	return err
+}
+
+// Recover takes a pointer to the returned error,
+// and sets it upon panic
+func Recover(err *error) {
+	if r := recover(); r != nil {
+		*err = NormalizePanic(r)
+	}
+}
+
+// WithType is a helper to augment an error with a corresponding type message
+func WithType(err error, obj interface{}) error {
+	return Wrap(err, fmt.Sprintf("%T", obj))
 }
