+++ errors/errors_test.go
@@ -254,3 +254,32 @@ func TestStackTrace(t *testing.T) {
 		})
 	}
 }
+
+// CheckErr is the type of all the check functions here
+type CheckErr func(error) bool
+
+// NoErr is useful for test cases when you want to fulfil the CheckErr type
+func NoErr(err error) bool {
+	return err == nil
+}
+
+// TestChecks make sure the Is and Err methods match
+func TestChecks(t *testing.T) {
+	cases := []struct {
+		err   error
+		check CheckErr
+		match bool
+	}{
+
+		// make sure lots of things match ErrInternal, but not everything
+		{Wrap(fmt.Errorf("internal"), "wrapped"),
+			func(err error) bool { return !Is(err, ErrInternal.New("wrapped")) }, true},
+		{nil, NoErr, true},
+		{Wrap(nil, "asd"), NoErr, true},
+	}
+
+	for i, tc := range cases {
+		match := tc.check(tc.err)
+		assert.Equal(t, tc.match, match, "%d", i)
+	}
+}
